# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Shipper list API",
        "version": "0.0.1"
    },
    "paths": {
...
        "/api/v1/shippers": {
            "get": {
                "tags": [
                    "Shipper"
                ],
                "summary": "Prints list of shippers",
                "operationId": "30982836d3e456a5efb03e9cf9c61e57",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filter by name",
                        "schema": {
                            "type": "string",
                            "example": "Kris Inc"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Shipper"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "422": {
                        "description": "Request has validation errors"
                    },
                    "429": {
                        "description": "Too many requests"
                    }
                }
            },
            "post": {
                "tags": [
                    "Shipper"
                ],
                "summary": "Create a new shipper record",
                "operationId": "ae4d9bbf1ca426010f6ceb33a5b8f691",
                "requestBody": {
                    "request": "ShipperCreateRequest",
                    "description": "Shipper data to be created",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Bills Inc"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "47 Upperthong Ln\\nHolmfirth\\nHD9 3UZ"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "422": {
                        "description": "Request has validation errors"
                    },
                    "429": {
                        "description": "Too many requests"
                    }
                },
                "security": {
                    "sanctum": {}
                }
            }
        },
        "/api/v1/shippers/{shipper-id}": {
            "get": {
                "tags": [
                    "Shipper"
                ],
                "summary": "Prints details for a single shipper",
                "operationId": "ce4e7eb03c06cc7236583374c8b9a7e5",
                "parameters": [
                    {
                        "name": "shiiper-id",
                        "in": "path",
                        "description": "Shipper ID",
                        "schema": {
                            "type": "integer",
                            "example": "6"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Shipper"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "422": {
                        "description": "Request has validation errors"
                    },
                    "429": {
                        "description": "Too many requests"
                    }
                }
            },
            "put": {
                "tags": [
                    "Shipper"
                ],
                "summary": "Update a shipper record",
                "operationId": "723ad850a6ecf8b4e5c4b56153e0b5c5",
                "requestBody": {
                    "request": "ShipperUpdateRequest",
                    "description": "Shipper data to be updated",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Bills Inc"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "47 Upperthong Ln\\nHolmfirth\\nHD9 3UZ"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "422": {
                        "description": "Request has validation errors"
                    },
                    "429": {
                        "description": "Too many requests"
                    }
                },
                "security": {
                    "sanctum": {}
                }
            },
            "delete": {
                "tags": [
                    "Shipper"
                ],
                "summary": "Delete a specified resource",
                "description": "Remove the specified resource from storage.",
                "operationId": "5bc464636a7500ba1c0696d57b0bc53e",
                "parameters": [
                    {
                        "name": "shipper_id",
                        "in": "path",
                        "description": "Shipper id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Shipper"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "422": {
                        "description": "Request has validation errors"
                    },
                    "429": {
                        "description": "Too many requests"
                    }
                },
                "security": {
                    "sanctum": {}
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Shipper": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Monster Inc."
                    },
                    "address": {
                        "type": "string",
                        "example": "722 Heidenreich Bridge\\nNorth Ursula, FL 04743"
                    },
                    "created_at": {
                        "type": "timestamp",
                        "example": "1970-01-01 00:00:00"
                    },
                    "updated_at": {
                        "type": "timestamp",
                        "example": "1970-01-01 00:00:00"
                    },
                    "deleted_at": {
                        "type": "timestamp",
                        "example": "1970-01-01 00:00:00"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Contact"
                }
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}